requirements.txt
ib_insync
psycopg2-binary
python-dotenv
pandas
numpy